% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.dragon_curve.R
\name{dragon_curve}
\alias{dragon_curve}
\title{Generate the Dragon Curve turn sequence}
\usage{
dragon_curve(iterations)
}
\arguments{
\item{-}{iterations An integer indicating the number of iterations
to be performed.}
}
\value{
A numeric vector of 0s and 1s representing left and right turns
in the Dragon Curve construction.
}
\description{
This function generates the sequence of turns corresponding to the
Dragon Curve fractal. Starting from an empty sequence, each iteration
appends a right turn (represented by 1) followed by the inverted and
reversed version of the previous sequence.
}
\details{
At each iteration, the sequence is updated according to the rule:
\deqn{S_{n+1} = S_n \; 1 \; \text{invert}(S_n)}
where \code{invert} denotes reversing the sequence and swapping
0s and 1s.
}
\examples{
dragon_curve(1)
dragon_curve(3)

}
\author{
linapeter
}
