% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.exercise_leapyears.R
\name{leap}
\alias{leap}
\title{Leap year}
\usage{
leap(year)
}
\arguments{
\item{-}{year Integer. A positive number indicating the year in question}
}
\value{
A logical value: \code{TRUE} if the year is a leap year,
\code{FALSE} otherwise.
}
\description{
This function returns whether the year is a leap year or not
}
\details{
This function corresponds to the Leap Year exercise from Exercism.
}
\examples{
leap(2000)
leap(1900)

}
\author{
linapeter
}
