% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.ducci_sequence.R
\name{step__}
\alias{step__}
\title{(Internal) Perform one step of the Diffy Game}
\usage{
step__(sequence)
}
\arguments{
\item{-}{sequence A numeric vector.}
}
\value{
A numeric vector of the same length as \code{sequence}, containing the
absolute differences.
}
\description{
Given a numeric sequence, this function computes the absolute differences
between consecutive elements, treating the sequence as circular (i.e.,
the last element is compared with the first).
}
\details{
For a sequence \eqn{(x_1, x_2, \dots, x_n)}, the result is
\eqn{(|x_1 - x_2|, |x_2 - x_3|, \dots, |x_n - x_1|)}.
}
\examples{
step__(c(1, 5, 3, 9))

}
\author{
linapeter
}
\keyword{internal}
