% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.exercise6_PE.R
\name{sum_squares_difference}
\alias{sum_squares_difference}
\title{sum_squares_difference}
\usage{
sum_squares_difference(limit)
}
\arguments{
\item{-}{limit Integer. A positive number indicating the upper bound of the
sequence starting at 1.}
}
\value{
A numeric value representing the difference between the square of the sum
and the sum of the squares from 1 to \code{limit}.
}
\description{
This function returns the difference between Sum of squares and Square of
sums, beginning on 1 until limit.
}
\details{
This function corresponds to Problem 6 of the Project Euler problem set.
}
\examples{
sum_squares_difference(10)
sum_squares_difference(100)

}
\author{
linapeter
}
