% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.ducci_sequence.R
\name{diffy_game}
\alias{diffy_game}
\title{Diffy Game for Multiple Sequences}
\usage{
diffy_game(sequences)
}
\arguments{
\item{sequences}{A list of numeric vectors (or objects coercible to numeric vectors),
each representing an initial sequence for the Diffy Game.}
}
\value{
A vector of integers, where each element corresponds to the number of
iterations performed for the respective input sequence.
}
\description{
Applies the Ducci (Diffy Game) process to multiple sequences.
Each sequence is processed independently using \code{ducci_sequence()}.
}
\details{
This function is a thin wrapper around \code{lapply()}, delegating the
computation to \code{ducci_sequence()} for each element of \code{sequences}.
}
\examples{
seqs <- list(
  c(1, 5, 7, 9, 9),
  c(1, 2, 1, 2, 1, 0),
  c(10, 12, 41, 62, 31, 50),
  c(10, 12, 41, 62, 31)
)

diffy_game(seqs)

}
\author{
linapeter
}
