% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.dragon_curve.R
\name{invert__}
\alias{invert__}
\title{(Internal) Invert and reverse a binary sequence}
\usage{
invert__(sequence)
}
\arguments{
\item{sequence}{A numeric vector containing only 0s and 1s.}
}
\value{
A numeric vector corresponding to the inverted and reversed input sequence.
}
\description{
This function takes a binary sequence (containing only 0s and 1s),
inverts its values (0 becomes 1 and 1 becomes 0), and then reverses
the order of the sequence.
}
\examples{
invert__(c(0, 1, 1, 0))

}
\author{
linapeter
}
\keyword{internal}
