% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.exercise8_PE.R
\name{largest_product}
\alias{largest_product}
\title{Largest product of consecutive digits}
\usage{
largest_product(digits)
}
\arguments{
\item{digits}{An integer indicating how many consecutive elements
are used to compute each product.}

\item{series}{A numeric vector containing the sequence of numbers.}
}
\value{
A numeric value corresponding to the largest product of
\code{digits} consecutive elements in the sequence.
}
\description{
This function finds the greatest product obtained from multiplying
a fixed number of consecutive elements in a numeric sequence.
}
\details{
This function corresponds to Exercise 8 from the Project Euler
problem set.
}
\examples{
largest_product(c(2, 3, 4, 3, 2, 5, 3, 2, 1, 4), 2)

}
\author{
Lina Park
}
